Add cached and in-progress response tracking
In the course of using this SDK I came upon use cases where I
was making multiple requests to the same dataset over and over again.
I implemented caching in the consumption of this SDK, but I also wanted
the SODA client to be smart about when it makes requests as sometimes
different objects were making the same request to the same SODA client.

This update ensures that if a request is already in progress to
a dataset, the completion handler provided will be cached and called
when the request completes.

It also offers an optional cache so that it will keep responses around
for a configurable amount of time if the same request is made more than once.

The caching functionality is enabled by default here, but may be turned off
by setting the TTL to 0.

The in progress request tracking may not be turned off.